Move audio and video information, and display decoders
